modularize scraper api controller
